fix: add upgrade test + fix
